rpm: Package (/var/)run/ceph
We currently do not package (/var/)run/ceph. This makes the directory
inherit the SELinux of the parent directory when it is created in the
post script instead of getting the proper context as defined in the
policy (which rpm handles on its own).

This follows the fedora packaging guideline on tmpfiles.d:

https://fedoraproject.org/wiki/Packaging:Tmpfiles.d

Signed-off-by: Boris Ranto <branto@redhat.com>