rpm: Do not pass interpret to postun scripts
If the interpret is passed to postun script, everything until the next
rpm section gets interpreted as the input for the interpret (in this
case ldconfig). This results in warnings like

/sbin/ldconfig: relative path `0' used to build cache

when removing the packages that pass the interpret because we use
several '#' characters to separate the subpackage sections and these get
passed to the ldconfig.

Using the default interpret (bash) fixes the warning. I'm also doing the
same to %post script for consistency.

Signed-off-by: Boris Ranto <branto@redhat.com>